+++ src/main/scala/com/databricks/labs/mosaic/core/types/model/CDMDataset.scala
@@ -1,0 +1,1 @@
+package com.databricks.labs.mosaic.core.types.model case class CDMDataset ()
