+++ src/main/scala/com/databricks/labs/mosaic/core/types/model/CDMStructure.scala
