+++ src/main/scala/com/databricks/labs/mosaic/expressions/format/CDMContent.scala
@@ -1,0 +1,1 @@
+package com.databricks.labs.mosaic.expressions.format case class CDMContent ()
